auth_service/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
django-cors-headers
drf-yasg
django-filter
uuid6










